# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation Taeyoung",
        "description": "API Documentation for Taeyoung",
        "version": "1.0.0"
    },
    "paths": {
...
        "/taeyoung-backend/public/api/quotation": {
            "get": {
                "tags": [
                    "Quotation"
                ],
                "summary": "Get all quotations",
                "description": "Get all quotations",
                "operationId": "b6ad0ebbf7b868424b0ad8064cd88114",
                "parameters": [
                    {
                        "parameter": "all",
                        "name": "all",
                        "in": "query",
                        "description": "Get all units",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "parameter": "page",
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "parameter": "per_page",
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "parameter": "sort",
                        "name": "sort",
                        "in": "query",
                        "description": "Sort by column",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "parameter": "direction",
                        "name": "direction",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "parameter": "number",
                        "name": "number",
                        "in": "query",
                        "description": "Filter by number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "parameter": "date",
                        "name": "date[]",
                        "in": "query",
                        "description": "Filter by date",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "parameter": "detail",
                        "name": "detail",
                        "in": "query",
                        "description": "Filter by detail",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "parameter": "client$filterName",
                        "name": "client$filterName",
                        "in": "query",
                        "description": "Filter by client name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Quotation"
                ],
                "summary": "Store a new quotation",
                "description": "Store a new quotation",
                "operationId": "a10f7bd838842d59e0ce4ef770879cfd",
                "requestBody": {
                    "description": "Quotation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreQuotationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuotationResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/taeyoung-backend/public/api/quotation/{id}": {
            "get": {
                "tags": [
                    "Quotation"
                ],
                "summary": "Show a quotation",
                "description": "Show a quotation",
                "operationId": "b0b55af43f7c86f02310f576a189dbc9",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "Quotation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuotationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quotation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Quotation not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Quotation"
                ],
                "summary": "Update a quotation",
                "description": "Update a quotation",
                "operationId": "c2c613321fabe55f3b62e22065107c59",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "Quotation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Quotation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateQuotationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuotationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quotation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Quotation not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Quotation"
                ],
                "summary": "Delete a quotation",
                "description": "Delete a quotation",
                "operationId": "ca0ae3d82574039bb8b63d269b0474bf",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "Quotation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Quotation deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quotation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Quotation not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Quotation cannot be deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Quotation cannot be deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "The pagination must be an integer."
                    }
                },
                "type": "object"
            },
            "Unauthenticated": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "StoreQuotationRequest": {
                "title": "StoreQuotationRequest",
                "required": [
                    "date",
                    "detail",
                    "discount",
                    "currencyType",
                    "client_id",
                    "detailMachinery",
                    "detailSpareParts"
                ],
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-08-19"
                    },
                    "detail": {
                        "type": "string",
                        "example": "This is a detail"
                    },
                    "discount": {
                        "type": "number",
                        "example": "0"
                    },
                    "currencyType": {
                        "type": "string",
                        "example": "USD"
                    },
                    "client_id": {
                        "type": "integer",
                        "example": "21"
                    },
                    "detailMachinery": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DetailMachineryRequest"
                        }
                    },
                    "detailSpareParts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DetailSparePartRequest"
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdateQuotationRequest": {
                "title": "UpdateQuotationRequest",
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-08-19"
                    },
                    "detail": {
                        "type": "string",
                        "example": "This is a detail"
                    },
                    "discount": {
                        "type": "number",
                        "example": "0"
                    },
                    "currencyType": {
                        "type": "string",
                        "example": "USD"
                    },
                    "client_id": {
                        "type": "integer",
                        "example": "21"
                    },
                    "detailMachinery": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DetailMachineryRequest"
                        }
                    },
                    "detailSpareParts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DetailSparePartRequest"
                        }
                    }
                },
                "type": "object"
            },
...
            "PurchaseCollection": {
                "title": "PurchaseCollection",
                "description": "Purchase resource collection",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PurchaseResource"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}